Python file used for data pre processing and model building
IDE used was Pycharm. Pandas, Matplotlib, Seaborn and Scikit learn libraries were used in this model building and data preprocessing tasks. Pandas was used for analysis and cleaning, Matplotlib and Seaborn for visually showing the model performance and Scikit-Learn for building the ML model- Decision Tree Classifier, and it's evaluation using multiple metrics.